bot/pyproject.toml
[tool.poetry]
name = "discordscrapebot"
version = "0.1.0"
description = "Bot to scrape discord messages from a server and store them to a database."
authors = ["Ethan Christensen <ethanlchristensen@outlook.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "DiscordScrapeBot"}]

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
numpy = "1.26.4"
pandas = "2.2.2"
discord = "^2.3.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
